refactor: breakchange directory structure
